requirements.txt
openai>=1.0.0
pydantic>=2.0.0
duckduckgo-search>=6.0.0
requests>=2.31.0
pyyaml>=6.0
rich>=13.0.0
